Refactorization
- admin user creation
- package hierarchy
- test endpoint can be accessed without authentication
