+++ src/main/java/ro/msg/learning/shop/security/SecurityConfig.java
@@ -29,6 +29,7 @@ public class SecurityConfig {
         http
                 .csrf(AbstractHttpConfigurer::disable)
                 .authorizeHttpRequests(authz -> authz
+                        .requestMatchers("/test/**").permitAll()
                         .anyRequest().authenticated()
                 );
