+++ src/main/java/ro/msg/learning/shop/service/CustomerService.java
@@ -1,26 +1,11 @@
 package ro.msg.learning.shop.service;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
 import ro.msg.learning.shop.model.Customer;
-import ro.msg.learning.shop.repository.CustomerRepository;
 
 import java.util.Optional;
 
-@Service
-@RequiredArgsConstructor
-public class CustomerService {
+public interface CustomerService {
+    Optional<Customer> findCustomerByUsername(String username);
 
-    private final CustomerRepository customerRepository;
-    private final PasswordEncoder passwordEncoder;
-
-    public Optional<Customer> findCustomerByUsername(String username) {
-        return customerRepository.findByUsername(username);
-    }
-
-    public Customer registerCustomer(Customer customer) {
-        customer.setPassword(passwordEncoder.encode(customer.getPassword()));
-        return customerRepository.save(customer);
-    }
+    void registerCustomer(Customer customer);
 }
