+++ src/main/java/ro/msg/learning/shop/service/CustomerServiceImpl.java
@@ -1,0 +1,28 @@
+package ro.msg.learning.shop.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+import ro.msg.learning.shop.model.Customer;
+import ro.msg.learning.shop.repository.CustomerRepository;
+
+import java.util.Optional;
+
+@Service
+@RequiredArgsConstructor
+public class CustomerServiceImpl implements CustomerService {
+
+    private final CustomerRepository customerRepository;
+    private final PasswordEncoder passwordEncoder;
+
+    @Override
+    public Optional<Customer> findCustomerByUsername(String username) {
+        return customerRepository.findByUsername(username);
+    }
+
+    @Override
+    public void registerCustomer(Customer customer) {
+        customer.setPassword(passwordEncoder.encode(customer.getPassword()));
+        customerRepository.save(customer);
+    }
+}
