+++ src/main/java/ro/msg/learning/shop/test/IntegrationTestDBService.java
@@ -1,0 +1,8 @@
+package ro.msg.learning.shop.test;
+
+import org.springframework.transaction.annotation.Transactional;
+
+public interface IntegrationTestDBService {
+    @Transactional
+    void clearAndPopulateDB();
+}
