+++ src/main/java/ro/msg/learning/shop/test/TestController.java
@@ -1,16 +1,15 @@
-package ro.msg.learning.shop.controller;
+package ro.msg.learning.shop.test;
 
 import lombok.AllArgsConstructor;
 import org.springframework.context.annotation.Profile;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import ro.msg.learning.shop.service.IntegrationTestDBService;
 
 @RestController
 @Profile("integrationtest")
 @AllArgsConstructor
-@RequestMapping("/db")
+@RequestMapping("/test")
 public class TestController {
     private final IntegrationTestDBService service;
