+++ src/main/resources/application.properties
@@ -9,3 +9,8 @@ ro.msg.learning.shop.locationselect.algorithm=PRIORITY_BASED
 ro.msg.learning.shop.locationselect.routematrix.apikey=${MAPQUEST_KEY}
 ro.msg.learning.shop.locationselect.routematrix.url=https://www.mapquestapi.com/directions/v2/routematrix
 ro.msg.learning.shop.security.authenticationType=HTTP_BASIC
+ro.msg.learning.shop.admin.firstname=${ADMIN_FIRST_NAME:admin}
+ro.msg.learning.shop.admin.lastname=${ADMIN_LAST_NAME:admin}
+ro.msg.learning.shop.admin.username=${ADMIN_USERNAME:admin}
+ro.msg.learning.shop.admin.password=${ADMIN_PASSWORD:admin}
+ro.msg.learning.shop.admin.email=${ADMIN_EMAIL:admin@admin.com}
